Move operator utility functions to /internal (#7)
This commit moves `types` and `utils` packages into folder `/internal`,
as it only contains code that performs operation that is useful only to
the operator itself and doesn't involve any other parties.
