+++ Dockerfile
@@ -11,10 +11,9 @@ RUN go mod download
 
 # Copy the go source
 COPY main.go main.go
-COPY types/ types/
 COPY controllers/ controllers/
-COPY servicedirectory/ servicedirectory/
-COPY utils/ utils/
+COPY pkg/ pkg/
+COPY internal/ internal/
 
 # Build
 RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 GO111MODULE=on go build -a -o manager *.go
