+++ controllers/service_controller.go
@@ -20,10 +20,10 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/CloudNativeSDWAN/cnwan-operator/internal/utils"
 	sr "github.com/CloudNativeSDWAN/cnwan-operator/pkg/servregistry"
-	"github.com/CloudNativeSDWAN/cnwan-operator/utils"
 
-	cnwan_types "github.com/CloudNativeSDWAN/cnwan-operator/types"
+	cnwan_types "github.com/CloudNativeSDWAN/cnwan-operator/internal/types"
 	"github.com/go-logr/logr"
 	"github.com/spf13/viper"
 	corev1 "k8s.io/api/core/v1"
